requirements.txt
python-telegram-bot==20.6
asyncio
logging
datetime
json
os
typing